requirements.txt
discord.py>=1.0.0
colorama>=0.3.9
requests>=2.5
weather-api>=0.0.6
